completed vehicle/user cant be deleted if booking is active

+++ src/modules/users/user.controller.ts
@@ -39,6 +39,15 @@ const updateUser = async (req: Request, res: Response) => {
 
 const deleteUser = async (req: Request, res: Response) => {
   try {
+    const isBookingExist = await userService.isBookingExist(
+      req.params.userId as string
+    );
+    if (isBookingExist) {
+      return res.status(400).json({
+        succes: false,
+        message: "User has Active vehicle Booking",
+      });
+    }
     const result = await userService.deleteUser(req.params.userId as string);
     if (result.rowCount === 0) {
       res.status(404).json({
